requirements.txt
pyrogram
requests
python-dotenv
openai
aiofiles
aiohttp